+++ flexiznam/main.py
@@ -1227,6 +1227,7 @@ def get_datasets(
     project_id=None,
     flexilims_session=None,
     filter_datasets=None,
+    exclude_datasets=None,
     allow_multiple=True,
     return_paths=False,
     return_dataseries=False,
@@ -1243,6 +1244,8 @@ def get_datasets(
             `flexilims_session` is provided.
         flexilims_session (:py:class:`flexilims.Flexilims`): Flexylims session object
         filter_datasets (dict): dictionary of key-value pairs to filter datasets by.
+        exclude_datasets (dict): dictionary of key-value pairs to exclude datasets by.
+            This acts as inverse filter.
         allow_multiple (bool): if True, allow multiple datasets to be returned,
             otherwise ensure that only one dataset exists online and return it.
         return_paths (bool): if True, return a list of paths
@@ -1271,6 +1274,14 @@ def get_datasets(
         flexilims_session=flexilims_session,
         filter=filter_datasets,
     )
+
+    if exclude_datasets is not None:
+        keep_dataset = pd.Series(True, index=datasets.index)
+        for key, value in exclude_datasets.items():
+            if key not in datasets.columns:
+                continue
+            keep_dataset &= datasets[key] != value
+        datasets = datasets[keep_dataset]
 
     if not return_dataseries:
         datasets = [