+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v0.4.3
 
+### Main changes
+
+- Add an option for inverse filtering in `get_datasets`. Use `exclude_datasets` to
+  filter out dataset with a specified value of an attribute.
+
 ### Bugfix
 
 - `extra_attributes` of datasets is now always a dictionary and cannot be set to `None`